# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Social",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/posts/": {
      "get": {
        "tags": [
          "Post"
        ],
        "description": "取得所有貼文",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "timeSort",
            "in": "query",
            "description": "時間排序，asc 為遞增，desc 為遞減",
            "required": false,
            "type": "string"
          },
          {
            "name": "q",
            "in": "query",
            "description": "搜尋關鍵字",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "string"
                          },
                          "content": {
                            "type": "string",
                            "example": "string"
                          },
                          "user": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string",
                                "example": "string"
                              },
                              "name": {
                                "type": "string",
                                "example": "string"
                              },
                              "photo": {
                                "type": "string",
                                "example": "string"
                              }
                            }
                          },
                          "commentsV": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "_id": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "comment": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "user": {
                                  "type": "object",
                                  "properties": {
                                    "_id": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "name": {
                                      "type": "string",
                                      "example": "string"
                                    }
                                  }
                                }
                              }
                            }
                          },
                          "createdAt": {
                            "type": "string",
                            "example": "string"
                          },
                          "updatedAt": {
                            "type": "string",
                            "example": "string"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "取得所有貼文成功"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized. 無效或缺少token。"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal server error."
          }
        }
      },
      "post": {
        "tags": [
          "Post"
        ],
        "description": "新增單筆貼文",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "新增貼文的資料",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "string"
                },
                "tags": {
                  "type": "array",
                  "example": [
                    "string"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "image": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "string"
                        },
                        "content": {
                          "type": "string",
                          "example": "string"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "example": "string"
                            },
                            "name": {
                              "type": "string",
                              "example": "string"
                            },
                            "photo": {
                              "type": "string",
                              "example": "string"
                            }
                          }
                        },
                        "tags": {
                          "type": "array",
                          "example": [
                            "string"
                          ],
                          "items": {
                            "type": "string"
                          }
                        },
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "image": {
                          "type": "string",
                          "example": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "string"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "新增成功"
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad request. 沒有填寫content資料或圖片網址格式不正確。"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized. 無效或缺少token。"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal server error."
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": [
          "Post"
        ],
        "description": "取得單一貼文",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文的ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "string"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "string"
                    },
                    "content": {
                      "type": "string",
                      "example": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "string"
                        },
                        "name": {
                          "type": "string",
                          "example": "string"
                        },
                        "photo": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "commentsV": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "string"
                          },
                          "comment": {
                            "type": "string",
                            "example": "string"
                          },
                          "user": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string",
                                "example": "string"
                              },
                              "name": {
                                "type": "string",
                                "example": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "string"
                    },
                    "updatedAt": {
                      "type": "string",
                      "example": "string"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "取得單一貼文成功"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized. 無效或缺少token。"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not found. 找不到該貼文。"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal server error."
          }
        }
      }
    },
    "/posts/posts": {
      "delete": {
        "tags": [
          "Post"
        ],
        "description": "刪除所有貼文",
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "string"
                },
                "data": {
                  "type": "array",
                  "example": [],
                  "items": {}
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "刪除所有貼文成功"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized. 無效或缺少token。"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal server error."
          }
        }
      }
    },
...
  }
}